Spring data JPA + H2 + MockMVC Test + @Query

+++ pocs/spring-boot-2x-data-jpa-h2/src/main/java/com/github/diegopacheco/sandboxspring/service/BookService.java
@@ -31,7 +31,12 @@ public class BookService {
     }
 
     public void deleteById(Long id){
-        dao.deleteById(id);
+        Boolean exist = dao.existsById(id);
+        if (exist){
+            dao.deleteById(id);
+            return;
+        }
+        throw new BookNotFoundException("Book with id: "+id+" Not Found. Impossible to delete record.");
     }
 
 }
